Cart

+++ model/cart.js
@@ -2,12 +2,8 @@ const db = require("../config/config");
 
 
 class Cart {
-    //Show all cart items
-    fetchItems(req, res) {
-        const query = `
-        SELECT  userID, cartID, prodName, quantity, amount, prodURL
-        FROM Cart;
-        `
+    fetchCart(req, res) {
+        const query = `SELECT Cart.cartID, Products.prodName AS prodName, Products.amount, Products.quantity, Products.prodUrl FROM Cart JOIN Products ON Cart.prodID = Products.prodID WHERE Cart.userID = ${req.params.id};`
         db.query(query, (err, results) => {
             if (err) throw err
             res.json({
@@ -16,49 +12,46 @@ class Cart {
             })
         })
     }
-    //Add to Cart
-    addItem(req, res) {
-        const query = `
-        INSERT INTO Cart
-        SET ?;
-        `
 
-        db.query(query, [req.body], (err) => {
+    addToCart(req, res) {
+        const query = `INSERT INTO Cart ?;`
+        db.query(query, data, (err) => {
             if (err) throw err
             res.json({
                 status: res.statusCode,
-                msg: "Item successfully added to cart"
+                msg: "Product has been successfully added"
             })
         })
     }
-    //Delete item from Cart
-    deleteItem(req, res) {
-        const query = `
-        DELETE FROM Cart
-        WHERE prodID = ${req.params.id};
-        `
+
+    //add something to cart
+    RemoveItem(req, res) {
+        const query = `DELETE FROM Cart WHERE cartID = ${req.params.id}`
         db.query(query, (err) => {
-            if (err) throw err,
-                res.json({
-                    status: res.statusCode,
-                    msg: "Item has been deleted"
-                })
+            if (err) throw err
+            res.json({
+                status: res.statusCode,
+                msg: "A product has been successfully deleted"
+            })
         })
     }
-    //delete specific item from cart
-        
-
-    //Edit existing product
-    updateItem(req, res) {
-        const query = `
-        UPDATE Cart
-        SET ? 
-        `
+    clearCart(req, res) {
+        const query = `DELETE FROM Cart WHERE userID = ?`
         db.query(query, [req.body, req.params.id], (err) => {
             if (err) throw err
             res.json({
                 status: res.statusCode,
-                msg: "Product has been updated"
+                msg: "Cart Cleared"
+            })
+        })
+    }
+    updateItems(req, res) {
+        const query = `UPDATE Cart SET ? WHERE cartID = ?;`
+        db.query(query, [req.body, req.params.id], (err) => {
+            if (err) throw err
+            res.json({
+                status: res.statusCode,
+                msg: "A product has been successfully updated"
             })
         })
     }
